Add Sherwin Williams color downloader
Sherwin Williams makes it difficult to download colors one at a time.
However, it's incredibly easy to download all colors in one go since
every time the web page displaying the details of one color is rendered,
the entire color catalog is downloaded to the client.  While inefficient
for regular users, it makes it easier to download colors for this brand.

Because of this fundamental change in behavior from how colors are
downloaded with Benjamin Moore, needed to refactor the download
interface to be more generic.  Rather than having "helpers", we have
"downloaders" now, which makes sense as more color brand downloaders are
added.

Functionality for downloading was more thoroughly encapsulated into
supporting classes.  This came as a by-product of refactoring to also
download Sherwin Williams colors.