Parallel download and create color swatches
Drop usage of the legacy WebClient class in favor of HttpClient even
though we are still using synchronous operations here, as there is no
need for asynchrony.

Want to make sure there is a single instance of HttpClient to avoid
creating more sockets than necessary and take advantage of connection
pooling.

At the moment, would rather feed all the inputs to a color downloader
that gives back swatches which can then be saved to a database.  The
alternative is to make a static HttpClient that is included as a member
of a color downloader from which downloaders for brands derive.  Then
it's reused everywhere, but disposal is more tricky when you have
multiple derivations floating around to know when the last one is used.
Some locking is also required here to ensure only one instance is
created.  This color downloader helps to ensure a single instance since
all downloading functionality is consolidated into one place.