Clean up code
ColorWriter has no need for instance variables, so make it a static
class.

ColorMatch just binds associated data together, so make it a struct.

Apply common naming conventions to private class variables.

Default to using the ToString method of ColorFilter to get its name,
rather than exposing it as a property, to match the number of other
places in the code where ToString is overridden and used frequently.