filter: Add intensity criteria
The intensity of a color may be loosely calculated by dividing its LRV
by its saturation.  Using LRV as a means of how bright a color appears
and saturation as how much color there is.  The higher the number, the
less intense.  The lower the number, the more intense.

Builders (e.g., Holt and Lennar) will tend to use colors of an intensity
somewhere between 190 and 250.  This is an effective means to filter
colors.