download: Avoid Task AggregateException
Unwrap this exception by getting the awaiter and getting the result.
This create a blocking approach to async methods.

Do not want to implement async throughout the code here because this is
the only action being done, to download data, as a part of the download
operation.  This is no other work to do.  Prefer to keep things simple.