download: Follow HttpClient best practices
While the existing implementation didn't abuse HttpClient by creating a
new one for each request, it was doing most work synchronously.  In
addition, while this is a console application, most usage of HttpClient
is done with web applications.  The goal here is to apply best practices
in the web world to the console world.  To do this, we need to create
and use a ServiceColleciton, which is given to us automatically in a web
application.

From there, we use dependency injection with IHttpClientFactory to pool
and reuse HttpMessageHandler.  Because IHttpClientFactory manages the
lifetimes of HttpMessageHandler, it allows HttpClient to respect DNS
changes.

We chain dependency injection at each step to work with typed services
everywhere, as opposed to using named services.  We first add HttpClient
for each commercial color provider and then the respective singleton
(which creates one instance of a service when it is first requested and
then reused all times thereafter).  Order is important here because we
don't want to create a service before all parts of it can be created.
This singleton contains all of the logic for downloading color data
specific to that commercial color provider.

Each specific color client inherits from a parent color client which
uses Polly to make network requests more fault tolerant.  Two policies
are used: a retry and a timeout.  When retrying, we don't use a fixed
interval because this could create times when many requests are all
tried again at the same time, leading to more failures.  Instead we
apply a random jitter that spreads the load more evenly.  The timeout
policy is basic and waits for 60 seconds for any connection to be made.
The timeout is the innermost policy that once connected is wrapped in
the retry policy.

Because of this inheritance, we can use polymorphism to build up a list
of tasks containing all of the clients that are needed to download data,
and data can be downloaded all at once.

From there, switch to making all download objects class instances rather
than static objects.  These are not utility classes and contain specific
per-instance data.