download: Account for more cases where bad data is received
Sometimes downloading colors from Benjamin Moore doesn't work as
expected.  If after resiliently downloading a web page we still get a
non-OK response status code, we need to give up and return null instead
of a valid string.

To make this work, we introduce nullable reference types.  From there,
we have all download operations populate a concurrent bag in parallel if
the download was successful so that no aggregation needs to be done with
LINQ anymore at the very end.  We're just left with a list of color
swatches that can be saved to the database.